\name{finemap-package}
\alias{finemap-package}
\alias{finemap}
\docType{package}
\title{
\packageTitle{finemap}
}
\description{
\packageDescription{finemap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{finemap}
\packageIndices{finemap}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{finemap}

Maintainer: \packageMaintainer{finemap}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
